all refatcored. new apihandlers.js replaces api.js filehandler.js takes file han dling functionality out of handlers.js# handlers.js now minimal. apiRequest renamed handleApirequest new handleFileRequest function took the functionality from hanlders.js uncommented favicon.ico
